Cleanup nodeId

+++ app/src/main/java/io/aether/android/screens/device/settings/FabricsViewModel.kt
@@ -72,7 +72,7 @@ constructor(
       _uiState.value = UiState.Loading
       try {
         val deviceCurrentFabricIndex =
-            clustersHelper.readCurrentFabricIndexAttribute(nodeId.toLong())
+            clustersHelper.readCurrentFabricIndexAttribute(nodeId)
         val controllerFabricIndex = chipClient.chipDeviceController.getFabricIndex()
         val currentFabricIndex = deviceCurrentFabricIndex ?: controllerFabricIndex
         if (fabricIndex == currentFabricIndex) {
@@ -80,7 +80,7 @@ constructor(
           _uiState.value = currentState
           return@launch
         }
-        clustersHelper.removeFabric(nodeId.toLong(), fabricIndex)
+        clustersHelper.removeFabric(nodeId, fabricIndex)
         _uiState.value = refreshFabrics(nodeId, currentState)
       } catch (e: Exception) {
         Timber.e(e, "removeFabric failed")
@@ -93,15 +93,15 @@ constructor(
 
   private suspend fun refreshFabrics(nodeId: NodeId, fallbackState: UiState): UiState {
     return try {
-      devicesRepository.getDeviceByNodeId(nodeId.toLong())
-      val fabrics = clustersHelper.readFabricsAttribute(nodeId.toLong())
-      val nocs = clustersHelper.readNOCsAttribute(nodeId.toLong())
+      devicesRepository.getDeviceByNodeId(nodeId)
+      val fabrics = clustersHelper.readFabricsAttribute(nodeId)
+      val nocs = clustersHelper.readNOCsAttribute(nodeId)
       if (fabrics == null || nocs == null) {
         if (fallbackState is UiState.Loaded) fallbackState
         else UiState.Error(R.string.controllers_offline)
       } else {
         val deviceCurrentFabricIndex =
-            clustersHelper.readCurrentFabricIndexAttribute(nodeId.toLong())
+          clustersHelper.readCurrentFabricIndexAttribute(nodeId)
         val controllerFabricIndex = chipClient.chipDeviceController.getFabricIndex()
         val currentFabricIndex = deviceCurrentFabricIndex ?: controllerFabricIndex
         val fabricsByIndex = fabrics.associateBy { it.fabricIndex }
